requirements.txt
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0
yt-dlp>=2023.12.30
python-dotenv>=1.0.0
async-timeout>=4.0.3
